pyproject.toml
[project]
name = "telegram-expense-tracker"
version = "0.1.0"
description = "Telegram bot for expense tracking with receipt OCR via GPT-5.4-mini"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-telegram-bot[webhooks]>=21.10",
    "openai>=1.70",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.15",
    "pydantic-settings>=2.8",
    "apscheduler>=3.11",
    "weasyprint>=63",
    "jinja2>=3.1",
    "aiosmtplib>=3.0",
    "aiohttp>=3.11",
    "pillow>=11.0",
    "psycopg2-binary>=2.9.11",
    "imagehash>=4.3.2",
    "python-multipart>=0.0.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
